Moves SQLAlchemy models to `models.py` module
